separating componentes
separating componentes
