+++ src/components/SavedElementShow.js
@@ -1,4 +1,7 @@
-function SavedElementShow({ item,handleRemoveSavedItem }) {
+import { useContext } from "react";
+import ItemContext from "../context";
+function SavedElementShow({ item }) {
+  const {handleRemoveSavedItem}=useContext(ItemContext)
   return (
     <>
       {item.title}
